% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_package_to_config.R
\name{add_package_to_config}
\alias{add_package_to_config}
\title{Add packages to your packages config file}
\usage{
add_package_to_config(
  package_name,
  packages_config_file = "./content/config/packages_to_load"
)
}
\arguments{
\item{package_name}{Character; name of the package you want to add}

\item{packages_config_file}{Character; path to your packages config file}
}
\description{
A function for adding a package to your existing packages config file
}
