% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_exercise.R
\name{build_exercise}
\alias{build_exercise}
\title{Build a exercise}
\usage{
build_exercise(exercise_path = NULL)
}
\arguments{
\item{exercises_home}{Character; path to the exercises folder}
}
\description{
Builds a exercise and moves their files into the respective folders
}
