% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_exercises.R
\name{build_exercises}
\alias{build_exercises}
\title{Build the exercises}
\usage{
build_exercises(exercises_home = "./content/exercises/", build_all = FALSE)
}
\arguments{
\item{exercises_home}{Character; path to the exercises folder}

\item{build_all}{Logical; defines whether all exercises are built
or only those that have been modified. Relies on a CSV file containing checksums that is
created the first time \code{build_sesssion()} was used; default is \code{FALSE}}
}
\description{
Builds the exercises and moves their files into the respective folders
}
