% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_session.R
\name{build_session}
\alias{build_session}
\title{Build a session}
\usage{
build_session(session_path = NULL)
}
\arguments{
\item{session_path}{Character; path to the session}
}
\description{
Build session slides and moves their files into the respective folders
}
