% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_sessions.R
\name{build_sessions}
\alias{build_sessions}
\title{Build the sessions}
\usage{
build_sessions(sessions_home = "./content/sessions/", build_all = FALSE)
}
\arguments{
\item{sessions_home}{Character; path to the sessions folder}

\item{build_all}{Logical; defines whether all sessions are build or only the
changed ones. Relies on a CSV file comprising checksums that is created the
first time \code{build_sesssion()} was used; default is \code{FALSE}}
}
\description{
Build the sessions and moves their files accordingly
}
