% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_workshop_skeleton.R
\name{create_workshop_skeleton}
\alias{create_workshop_skeleton}
\title{Create workshop skeleton}
\usage{
create_workshop_skeleton(where = ".", style = "gesis")
}
\arguments{
\item{where}{Character; path to the workshop folder}

\item{style}{Character; currently only \code{gesis} is supported}
}
\description{
Takes care of creating all folder and subfolder for a workshop
}
