% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_picture.R
\name{include_picture}
\alias{include_picture}
\title{Add a picture to your xaringan presentation}
\usage{
include_picture(picture_name, picture_path = "./content/img/")
}
\arguments{
\item{picture_name}{Character; the name of the picture and its extension}

\item{picture_path}{Character; the path to the folder containing your pictures}
}
\description{
A path normalizing wrapper around code{knitr::include_graphics()}
}
